+++ internals.js
@@ -1,7 +1,0 @@
-window.internals = {
-    url: require('./src/url'),
-    config: require('./src/config'),
-    util: require('./src/util'),
-    grid: require('./src/grid'),
-    request: require('./src/request')
-};