+++ src/config.js
@@ -1,14 +1,8 @@
 'use strict';
 
 module.exports = {
-
-    HTTP_URLS: [
-        'http://a.tiles.mapbox.com/v3/',
-        'http://b.tiles.mapbox.com/v3/'],
-
+    HTTP_URL: 'http://a.tiles.mapbox.com/v4',
+    HTTPS_URL: 'https://a.tiles.mapbox.com/v4',
     FORCE_HTTPS: false,
-
-    HTTPS_URLS: [
-        'https://a.tiles.mapbox.com/v3/',
-        'https://b.tiles.mapbox.com/v3/']
+    REQUIRE_ACCESS_TOKEN: true
 };