+++ test/spec/marker.js
@@ -1,4 +1,14 @@
 describe('L.mapbox.marker', function() {
+    var retina = L.Browser.retina;
+
+    beforeEach(function() {
+        L.Browser.retina = false;
+    });
+
+    afterEach(function() {
+        L.Browser.retina = retina;
+    });
+
     describe('#style', function() {
         it("produces a small marker", function() {
             var marker = L.mapbox.marker.style({
@@ -6,7 +16,17 @@ describe('L.mapbox.marker', function() {
                     'marker-size': 'small'
                 }
             });
-            expect(marker.options.icon.options.iconUrl).to.contain('pin-s');
+            expect(marker.options.icon.options.iconUrl).to.equal(internals.url('/marker/pin-s+7e7e7e.png'));
+        });
+
+        it("uses @2x suffix on retina", function() {
+            L.Browser.retina = true;
+            var marker = L.mapbox.marker.style({
+                properties: {
+                    'marker-size': 'small'
+                }
+            });
+            expect(marker.options.icon.options.iconUrl).to.equal(internals.url('/marker/pin-s+7e7e7e@2x.png'));
         });
 
         it("produces a medium marker", function() {
@@ -25,6 +45,17 @@ describe('L.mapbox.marker', function() {
                 }
             });
             expect(marker.options.icon.options.iconUrl).to.contain('f00');
+        });
+
+        it("supports custom access token", function() {
+            var marker = L.mapbox.marker.style({
+                properties: {
+                    'marker-size': 'small'
+                }
+            }, [0, 0], {
+                accessToken: 'custom'
+            });
+            expect(marker.options.icon.options.iconUrl).to.equal(internals.url('/marker/pin-s+7e7e7e.png', 'custom'));
         });
 
         it("tolerates empty input", function() {
@@ -64,7 +95,16 @@ describe('L.mapbox.marker', function() {
             var icon = L.mapbox.marker.icon({
                 'marker-size': 'large'
             });
-            expect(icon.options.iconUrl).to.contain('pin-l');
+            expect(icon.options.iconUrl).to.equal(internals.url('/marker/pin-l+7e7e7e.png'));
+        });
+
+        it("supports custom access token", function() {
+            var icon = L.mapbox.marker.icon({
+                'marker-size': 'large'
+            }, {
+                accessToken: 'custom'
+            });
+            expect(icon.options.iconUrl).to.equal(internals.url('/marker/pin-l+7e7e7e.png', 'custom'));
         });
     });