+++ test/spec/tile_layer.js
@@ -33,7 +33,7 @@ describe("L.mapbox.tileLayer", function() {
         });
 
         it('loads TileJSON from a URL', function(done) {
-            var layer = L.mapbox.tileLayer('http://a.tiles.mapbox.com/v3/L.mapbox.map-0l53fhk2.json');
+            var layer = L.mapbox.tileLayer('http://a.tiles.mapbox.com/v3/mapbox.map-0l53fhk2.json');
 
             layer.on('ready', function() {
                 expect(this).to.equal(layer);
@@ -41,13 +41,13 @@ describe("L.mapbox.tileLayer", function() {
                 done();
             });
 
-            server.respondWith("GET", "http://a.tiles.mapbox.com/v3/L.mapbox.map-0l53fhk2.json",
+            server.respondWith("GET", "http://a.tiles.mapbox.com/v3/mapbox.map-0l53fhk2.json",
                 [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.tileJSON)]);
             server.respond();
         });
 
         it('loads TileJSON from an ID', function(done) {
-            var layer = L.mapbox.tileLayer('L.mapbox.map-0l53fhk2');
+            var layer = L.mapbox.tileLayer('mapbox.map-0l53fhk2');
 
             layer.on('ready', function() {
                 expect(this).to.equal(layer);
@@ -55,13 +55,27 @@ describe("L.mapbox.tileLayer", function() {
                 done();
             });
 
-            server.respondWith("GET", "http://a.tiles.mapbox.com/v3/L.mapbox.map-0l53fhk2.json",
+            server.respondWith("GET", internals.url.tileJSON('mapbox.map-0l53fhk2'),
+                [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.tileJSON)]);
+            server.respond();
+        });
+
+        it('supports custom access token', function(done) {
+            var layer = L.mapbox.tileLayer('mapbox.map-0l53fhk2', {accessToken: 'custom'});
+
+            layer.on('ready', function() {
+                expect(this).to.equal(layer);
+                expect(layer.getTileJSON()).to.eql(helpers.tileJSON);
+                done();
+            });
+
+            server.respondWith("GET", internals.url.tileJSON('mapbox.map-0l53fhk2', 'custom'),
                 [200, { "Content-Type": "application/json" }, JSON.stringify(helpers.tileJSON)]);
             server.respond();
         });
 
         it('emits an error event', function(done) {
-            var layer = L.mapbox.tileLayer('http://a.tiles.mapbox.com/v3/L.mapbox.map-0l53fhk2.json');
+            var layer = L.mapbox.tileLayer('mapbox.map-0l53fhk2');
 
             layer.on('error', function(e) {
                 expect(this).to.equal(layer);
@@ -69,7 +83,7 @@ describe("L.mapbox.tileLayer", function() {
                 done();
             });
 
-            server.respondWith("GET", "http://a.tiles.mapbox.com/v3/L.mapbox.map-0l53fhk2.json",
+            server.respondWith("GET", internals.url.tileJSON('mapbox.map-0l53fhk2'),
                 [400, { "Content-Type": "application/json" }, JSON.stringify({error: 'error'})]);
             server.respond();
         });
@@ -83,7 +97,17 @@ describe("L.mapbox.tileLayer", function() {
     });
 
     describe("#getTileUrl", function() {
-        beforeEach(setRetina(false));
+        var retina;
+
+        beforeEach(function() {
+            retina = L.Browser.retina;
+            L.Browser.retina = false;
+        });
+
+        afterEach(function() {
+            L.Browser.retina = retina;
+        });
+
         it("distributes over the URLs in the tiles property", function() {
             var layer = L.mapbox.tileLayer(helpers.tileJSON);
             expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/examples.map-8ced9urs/0/0/0.png');
@@ -92,6 +116,7 @@ describe("L.mapbox.tileLayer", function() {
             expect(layer.getTileUrl({x: 3, y: 0, z: 0})).to.equal('http://d.tiles.mapbox.com/v3/examples.map-8ced9urs/0/3/0.png');
             expect(layer.getTileUrl({x: 4, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/examples.map-8ced9urs/0/4/0.png');
         });
+
         it("changes format of tiles", function() {
             var layer = L.mapbox.tileLayer(helpers.tileJSON).setFormat('jpg70');
             expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/examples.map-8ced9urs/0/0/0.jpg70');
@@ -100,37 +125,11 @@ describe("L.mapbox.tileLayer", function() {
             expect(layer.getTileUrl({x: 3, y: 0, z: 0})).to.equal('http://d.tiles.mapbox.com/v3/examples.map-8ced9urs/0/3/0.jpg70');
             expect(layer.getTileUrl({x: 4, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/examples.map-8ced9urs/0/4/0.jpg70');
         });
-    });
 
-    describe("#autoScale", function() {
-        it("uses retina automatically", function() {
-            setRetina(true)();
-            var layer = L.mapbox.tileLayer(helpers.tileJSON_autoscale, {
-                detectRetina: true
-            });
-            expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/tmcw.map-oitj0si5/0/0/0@2x.png');
-        });
-        it("does not engage on non-retina systems", function() {
-            setRetina(false)();
-            var layer = L.mapbox.tileLayer(helpers.tileJSON_autoscale, {
-                detectRetina: true
-            });
-            expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/tmcw.map-oitj0si5/0/0/0.png');
-        });
-        it("does not engage with detectRetina: false", function() {
-            setRetina(true)();
-            var layer = L.mapbox.tileLayer(helpers.tileJSON_autoscale, {
-                detectRetina: false
-            });
-            expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/tmcw.map-oitj0si5/0/0/0.png');
+        it("requests @2x tiles on retina", function() {
+            L.Browser.retina = true;
+            var layer = L.mapbox.tileLayer(helpers.tileJSON);
+            expect(layer.getTileUrl({x: 0, y: 0, z: 0})).to.equal('http://a.tiles.mapbox.com/v3/examples.map-8ced9urs/0/0/0@2x.png');
         });
     });
-
-
-
-    function setRetina(x) {
-        return function() {
-            L.Browser.retina = x;
-        };
-    }
 });